Changed .int() to .long().

+++ src/losses/coherence.py
@@ -18,7 +18,7 @@ def calc_coherence_loss(agent_level, matrices, mask, eos_positions, embeddings):
 
     # todo: make sure the pad token is not here, also no join for levels 0 and 1 otherwise pad learns
     #random_indexes = torch.fmod(torch.randperm(batch * seq_length).to(Config.device), embeddings.shape[0])
-    random_indexes = (torch.rand(batch * seq_length).to(Config.device) * embeddings.shape[0]).floor().int()
+    random_indexes = (torch.rand(batch * seq_length).to(Config.device) * embeddings.shape[0]).floor().long()
     random_vec_replacements = torch.index_select(embeddings, 0, random_indexes)
     random_vec_replacements = random_vec_replacements.view(batch, seq_length, vec_size)
